backend/approvals/requirements.txt
pyjwt==2.8.0
psycopg2-binary==2.9.9
pydantic==2.5.0
